requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
streamlit>=1.12.0
googlemaps>=4.0.0
python-dotenv>=0.19.0
pydantic>=2.0.0
httpx>=0.24.0
python-multipart>=0.0.5
numpy>=1.21.0,<2.0.0
requests>=2.28.0
transformers>=4.21.0
torch>=1.12.0